pyproject.toml
[project]
name = "document-agent"
version = "0.1.0"
description = "A general-purpose AI assistant for document editing, research, and writing tasks"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.54.0",
    "logfire>=3.21.1",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
